moved delete file envelope and delete strategy to their own modules
